refactor: use nominal ClassDesc in AtomicOp.InstanceOf
Replaces the Class payload of InstanceOf with a ClassDesc, converted at
the construction sites in Lowering and SpecializeAndLower. Codegen emits
INSTANCEOF from the descriptor directly, deleting a JvmName.ofClass call
site. Introduces ClassDescs.of as the shared Class-to-ClassDesc
conversion, now also used by JField.

Part of #13091.
